ai-interior-style-back-end/ai_service/requirements.txt
fastapi
uvicorn
python-multipart
numpy
pillow
faiss-cpu
git+https://github.com/openai/CLIP.git
torch
torchvision
